requirements.txt
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
botocore==1.33.11
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
docutils==0.16
exceptiongroup==1.2.0
Flask==3.0.0
h11==0.14.0
idna==3.6
importlib-metadata==6.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
outcome==1.3.0.post0
pyasn1==0.5.1
PySocks==1.7.1
python-consul2==0.1.5
python-dateutil==2.8.2
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.8.2
schedule-sdk==3.0.7
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==1.26.18
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0

snowflake~=0.5.1
PyMySQL~=1.1.0
injector~=0.21.0